Fixed a bug where StaticRectangles were behaving like... lines

+++ server/Stage.cpp
@@ -46,7 +46,7 @@ Stage::Stage(size_t width, size_t height):
     this->timeStamp = std::chrono::system_clock::now();
 }
 
-b2Body* Stage::addStaticRectangle(size_t v_side, size_t h_side,
+b2Body* Stage::addStaticRectangle(float v_side, float h_side,
                                   float x_pos, float y_pos) {
     b2BodyDef body;
     body.type = b2_staticBody;
@@ -98,7 +98,7 @@ void Stage::addBrickBlock(size_t side, float x_pos, float y_pos) {
     brick_blocks.insert({coordinates, block});
 }
 
-void Stage::addMetalBlock(size_t side, float x_pos, float y_pos) {
+void Stage::addMetalBlock(float side, float x_pos, float y_pos) {
     if (x_pos < 0 || x_pos > width || y_pos < 0 || y_pos > height) {
         throw StageOutOfRangeException();
     }
